refactor: Colocate Explore page related files with Posts
