+++ src/features/core/api/loaders.ts
@@ -1,5 +1,5 @@
 import { QueryClient } from "@tanstack/react-query";
-import { currentUserQuery } from "../utils/auth";
+import { currentUserQuery } from "../../accounts/api/queries";
 
 export const appLoader = (queryClient: QueryClient) => async () => {
   const query = currentUserQuery();
