+++ src/features/core/hooks/useCurrentUserQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { currentUserQuery } from "../utils/auth";
+import { currentUserQuery } from "../../accounts/api/queries";
 
 export default function useCurrentUserQuery() {
   return useQuery(currentUserQuery());
