+++ src/features/posts/hooks/useAllPostsQuery.ts
